requirements.txt
langchain==0.3.9
langchain-openai==0.2.11
python-dotenv==1.0.0
openai==1.56.2
chromadb==0.5.21
tiktoken==0.8.0 
langchain-community==0.3.9